% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data by Group}
\usage{
summarize_data(data, group_var, summary_var, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame to be summarized.}

\item{group_var}{A string specifying the column name to group by.}

\item{summary_var}{A string specifying the numeric column to summarize.}

\item{na.rm}{A logical value indicating whether to remove NA values. Defaults to TRUE.}
}
\value{
A data frame with the grouping variable and the calculated mean and standard deviation.
}
\description{
This function takes a data frame, groups it by a specified column, and summarizes a numeric column by calculating the mean and standard deviation.
}
\examples{
summarize_data(mtcars, "cyl", "mpg")

# Handling NA values
mtcars_na <- mtcars
mtcars_na$mpg[c(1, 5, 10)] <- NA
summarize_data(mtcars_na, "cyl", "mpg", na.rm = TRUE)
summarize_data(mtcars_na, "cyl", "mpg", na.rm = FALSE)

}
